backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]
